Added Army Movement
Added values for the movement of armies on the world map based on the
speed of the contained units.
Added a check method to find the set of grid squares the selected army
can move to and also to check if the mouse is within one of those
squares.
Altered the paint method to highlight the grid squares the army can move
into blue and if the mouse lies within one of those squares then
highlight that square yellow