Added AI in the real time battles and army player ownership
Added a display in the army information panel in the over world map for
the player that owns that unit.
Added methods for AI targeting for the enemy units in a battle. Now at
intervals they calculate the best target from the list of units the user
controls and move towards that units location