Added location checking to army merges
Added a recursive function to relocate units when two armies are joined
so they are not colliding in the battle.
Fixed a minor issue with merging armies setting the hovering ID to the
wrong value.
Changed the real time battles to be paused at the start (The buttons for
game speed need to be changed to reflect this however)