Added creating new accounts
Filled in the method to create a new user account. This will add a new
user to the account table in the database with default armies and
settlement arrays. In doing so I also added a method in the save manager
that will save default values to whatever account name is given to it,
this will reset pre-existing files with the default values or create new
files for a new user account