Added checks to signing in
Added checks to make sure that the user signs in to an account when the
game loads and added the ability for the user to add a password to a
newly created account. Added checks on sign in to make sure both a
username and password are entered and then check if the account is in
the database and if the password was correct. Need to fill in the
methods to handle the encryption and decryption of the passwords