Added password encryption and minor fixes (Code Complete)
Added the methods so that the passwords entered by the user are
encrypted before being checked or added to the database. Also fixed the
issue with total settlement income not being added correctly to the
player funds.
Added a validation check so when building a unit the player must have
the funds necessary to buy one rather than allowing the player funds to
become negative