Fixed minor bugs
Fixed a bug where the movement distance of an army would not be
decreased if it was moved only vertically.
Fixed the issue where if an army was moved on top of a settlement and
then deselected then it could not be reselected and so moved of the
settlement.
Fixed a bug where when a building is removed from a settlement then the
income of the current building selected rather than the building being
removed is taken away from the settlement income